Consolidate 7 E2E verification test files into verification-core.spec.ts
Merge assertions from:
- verification-sessions.spec.ts (WS event shape, step definitions, active verifications API)
- verification-sessions-heavy.spec.ts (multi-step events, LLM-review sessionId, event ordering, auto-pass)
- verification-output.spec.ts (startedAt timestamps, step_output events)
- verification-output-multi.spec.ts (multi-step output fields, timestamp consistency)
- verification-modal-output.spec.ts (active verification API returns step output)
- error-pattern-verification.spec.ts (expect:failure pipeline integration)
- llm-review-verification.spec.ts (LLM review skip path)

Organized into 6 describe blocks:
1. Verification WS events (5 tests)
2. Multi-step and heavy verification (4 tests)
3. Verification output streaming (5 tests)
4. Active verification API (2 tests)
5. Expect failure pipeline (2 tests)
6. LLM Review verification (1 test)

All 19 tests pass. Files kept unchanged:
- cancel-verification.spec.ts
- verification-result.spec.ts
- ui/cancel-verification.spec.ts
- ui/gate-verification-ux.spec.ts
