--- tests/e2e/verification-output.spec.ts
@@ -1,148 +1,0 @@
-/**
- * E2E tests for live verification UX features:
- *
- * - gate_verification_started WS events include startedAt field
- * - gate_verification_step_started WS events include startedAt field
- * - gate_verification_step_output WS events are broadcast during command verification
- */
-import { test, expect } from "./in-process-harness.js";
-import {
-	apiFetch,
-	createGoal,
-	deleteGoal,
-	connectWs,
-	createSession,
-	deleteSession,
-} from "./e2e-setup.js";
-
-/** Create a goal using the test-fast workflow (command-only steps, fast). */
-async function createTestFastGoal(): Promise<string> {
-	const goal = await createGoal({ title: `Verification Output E2E ${Date.now()}`, workflowId: "test-fast" });
-	return goal.id;
-}
-
-/** Poll until a gate reaches the target status or timeout expires. */
-async function waitForGateStatus(
-	goalId: string,
-	gateId: string,
-	targetStatus: string,
-	timeoutMs = 30_000,
-): Promise<any> {
-	const start = Date.now();
-	while (Date.now() - start < timeoutMs) {
-		const res = await apiFetch(`/api/goals/${goalId}/gates/${gateId}`);
-		const data = await res.json();
-		if (data.status === targetStatus) return data;
-		await new Promise(r => setTimeout(r, 50));
-	}
-	throw new Error(`Gate ${gateId} did not reach "${targetStatus}" within ${timeoutMs}ms`);
-}
-
-test.describe("Verification output streaming and timestamps", () => {
-
-	test("gate_verification_started includes startedAt timestamp", async () => {
-		const goalId = await createTestFastGoal();
-		const sessionId = await createSession({ goalId });
-		const ws = await connectWs(sessionId);
-		try {
-			const before = Date.now();
-
-			await apiFetch(`/api/goals/${goalId}/gates/design-doc/signal`, {
-				method: "POST",
-				body: JSON.stringify({ content: "# Design\n\nTest content" }),
-			});
-
-			const started = await ws.waitFor(
-				(m) => m.type === "gate_verification_started" && m.gateId === "design-doc",
-				30_000,
-			);
-
-			const after = Date.now();
-
-			// startedAt must be present and a number
-			expect(typeof started.startedAt).toBe("number");
-			// Must be a reasonable timestamp (between before and after the signal)
-			expect(started.startedAt).toBeGreaterThanOrEqual(before);
-			expect(started.startedAt).toBeLessThanOrEqual(after);
-
-			await waitForGateStatus(goalId, "design-doc", "passed");
-		} finally {
-			ws.close();
-			await deleteSession(sessionId);
-			await deleteGoal(goalId);
-		}
-	});
-
-	test("gate_verification_step_started includes startedAt timestamp", async () => {
-		const goalId = await createTestFastGoal();
-		const sessionId = await createSession({ goalId });
-		const ws = await connectWs(sessionId);
-		try {
-			const before = Date.now();
-
-			await apiFetch(`/api/goals/${goalId}/gates/design-doc/signal`, {
-				method: "POST",
-				body: JSON.stringify({ content: "# Design\n\nTest content" }),
-			});
-
-			const stepStarted = await ws.waitFor(
-				(m) => m.type === "gate_verification_step_started" && m.gateId === "design-doc",
-				30_000,
-			);
-
-			const after = Date.now();
-
-			// startedAt must be present and a number
-			expect(typeof stepStarted.startedAt).toBe("number");
-			expect(stepStarted.startedAt).toBeGreaterThanOrEqual(before);
-			expect(stepStarted.startedAt).toBeLessThanOrEqual(after);
-			// Also has standard fields
-			expect(stepStarted.goalId).toBe(goalId);
-			expect(stepStarted.signalId).toBeTruthy();
-			expect(typeof stepStarted.stepIndex).toBe("number");
-			expect(stepStarted.stepName).toBe("Content present");
-
-			await waitForGateStatus(goalId, "design-doc", "passed");
-		} finally {
-			ws.close();
-			await deleteSession(sessionId);
-			await deleteGoal(goalId);
-		}
-	});
-
-	test("gate_verification_step_output events are broadcast for command steps", async () => {
-		const goalId = await createTestFastGoal();
-		const sessionId = await createSession({ goalId });
-		const ws = await connectWs(sessionId);
-		try {
-			await apiFetch(`/api/goals/${goalId}/gates/design-doc/signal`, {
-				method: "POST",
-				body: JSON.stringify({ content: "# Design\n\nTest" }),
-			});
-
-			// The test-fast design-doc gate runs "echo ok", so we should get stdout output.
-			// Use a generous timeout — under system load, child process spawning can be slow.
-			const output = await ws.waitFor(
-				(m) => m.type === "gate_verification_step_output" && m.gateId === "design-doc",
-				30_000,
-			);
-
-			// Validate all required fields
-			expect(output.goalId).toBe(goalId);
-			expect(output.gateId).toBe("design-doc");
-			expect(output.signalId).toBeTruthy();
-			expect(typeof output.stepIndex).toBe("number");
-			expect(output.stepIndex).toBe(0);
-			expect(output.stream).toBe("stdout");
-			expect(typeof output.text).toBe("string");
-			expect(output.text).toContain("ok");
-			expect(typeof output.ts).toBe("number");
-
-			await waitForGateStatus(goalId, "design-doc", "passed");
-		} finally {
-			ws.close();
-			await deleteSession(sessionId);
-			await deleteGoal(goalId);
-		}
-	});
-});
